requirements.txt
websockets>=12.0
python-dotenv>=1.0.0
pexpect>=4.9.0
